refactor(frontend): change name of DAOSection directory
